add initial implementation of RingBuffer
The ring buffer is one of the core data structures used in kobolds.
There are obvious improvements to make after this version like extending
the functionality. Additionally, there are some minor inconsitencies in
the function signatures as some return errors, other void, and others
u32s. I also think that we may want to change the return u32s to
something different in the future.